# Speedup sorting for inline views: 1.4x - 1.7x improvement