--- a/arrow-ord/src/sort.rs
+++ b/arrow-ord/src/sort.rs
@@ -316,8 +316,7 @@ fn sort_byte_view<T: ByteViewType>(
         .map(|idx| {
             // SAFETY: we know idx < values.len()
             let raw = unsafe { *values.views().get_unchecked(idx as usize) };
-            let len = raw as u32; // lower 32 bits encode length
-            (idx, raw, len)
+            (idx, raw)
         })
         .collect();
 
@@ -328,39 +327,27 @@ fn sort_byte_view<T: ByteViewType>(
     };
 
     // 3. Mixed comparator: first prefix, then inline vs full comparison
-    let cmp_mixed = |a: &(u32, u128, u32), b: &(u32, u128, u32)| {
-        let (_, raw_a, len_a) = *a;
-        let (_, raw_b, len_b) = *b;
-
+    let cmp_mixed = |a: &(u32, u128), b: &(u32, u128)| {
+        let (_, raw_a) = *a;
+        let (_, raw_b) = *b;
+        let len_a = raw_a as u32;
+        let len_b = raw_b as u32;
         // 3.1 Both inline (≤12 bytes): compare full 128-bit key including length
         if len_a <= MAX_INLINE_VIEW_LEN && len_b <= MAX_INLINE_VIEW_LEN {
             return GenericByteViewArray::<T>::inline_key_fast(raw_a)
                 .cmp(&GenericByteViewArray::<T>::inline_key_fast(raw_b));
         }
 
-        let l_byte_view = ByteView::from(raw_a);
-        let r_byte_view = ByteView::from(raw_b);
-
         // 3.2 Compare 4-byte prefix in big-endian order
-        let pref_a = l_byte_view.prefix.swap_bytes();
-        let pref_b = r_byte_view.prefix.swap_bytes();
+        let pref_a = ByteView::from(raw_a).prefix.swap_bytes();
+        let pref_b = ByteView::from(raw_b).prefix.swap_bytes();
         if pref_a != pref_b {
             return pref_a.cmp(&pref_b);
         }
 
         // 3.3 Fallback to full byte-slice comparison
-        // we can skip the first 4 bytes of the view, those are known to be equal
-        let data = unsafe { values
-            .data_buffers()
-            .get_unchecked(l_byte_view.buffer_index as usize) };
-        let offset = l_byte_view.offset as usize;
-        let full_a = unsafe { data.get_unchecked(offset + 4..offset + l_byte_view.length as usize) };
-
-        let data = unsafe { values
-            .data_buffers()
-            .get_unchecked(r_byte_view.buffer_index as usize) };
-        let offset = r_byte_view.offset as usize;
-        let full_b = unsafe { data.get_unchecked(offset + 4..offset + r_byte_view.length as usize) };
+        let full_a: &[u8] = unsafe { values.value_unchecked(a.0 as usize).as_ref() };
+        let full_b: &[u8] = unsafe { values.value_unchecked(b.0 as usize).as_ref() };
         full_a.cmp(full_b)
     };
 
@@ -380,10 +367,10 @@ fn sort_byte_view<T: ByteViewType>(
         // Place null indices first
         out.extend_from_slice(&nulls[..nulls.len().min(out_limit)]);
         let rem = out_limit - out.len();
-        out.extend(valids.iter().map(|&(i, _, _)| i).take(rem));
+        out.extend(valids.iter().map(|&(i, _)| i).take(rem));
     } else {
         // Place non-null indices first
-        out.extend(valids.iter().map(|&(i, _, _)| i).take(out_limit));
+        out.extend(valids.iter().map(|&(i, _)| i).take(out_limit));
         let rem = out_limit - out.len();
         out.extend_from_slice(&nulls[..rem]);
     }