# Which issue does this PR close?
- Closes #7857

# Rationale for this change

```
sort string_view[0-400] nulls to indices 2^12                                      1.00     45.2±1.37µs        ? ?/sec    1.01     45.8±1.74µs        ? ?/sec
sort string_view[0-400] to indices 2^12                                            1.00     69.1±1.98µs        ? ?/sec    1.00     69.1±4.24µs        ? ?/sec
sort string_view[10] nulls to indices 2^12                                         1.00     40.8±1.81µs        ? ?/sec    1.37     55.7±3.90µs        ? ?/sec
sort string_view[10] to indices 2^12                                               1.00     52.8±0.35µs        ? ?/sec    1.63     85.9±1.46µs        ? ?/sec
sort string_view_inlined[0-12] nulls to indices 2^12                               1.00     40.9±1.99µs        ? ?/sec    1.29     52.6±1.76µs        ? ?/sec
sort string_view_inlined[0-12] to indices 2^12                                     1.00     50.6±0.27µs        ? ?/sec    1.68    85.0±12.24µs        ? ?/sec
```
# What changes are included in this PR?

Speedup by specializing on batches with only inline views.

# Are these changes tested?, are they covered by existing tests)?

existing tests

# Are there any user-facing changes?

no